back/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

boto3==1.28.61
botocore==1.31.61
fastapi==0.103.1
pydantic==2.3.0
pydub==0.25.1
pymongo==4.5.0
PyPDF2==3.0.1
python-dotenv==1.0.0
starlette==0.27.0
uvicorn==0.23.2
Jinja2==3.1.2
python-multipart==0.0.6
